python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "web-crawler-sdk"
version = "0.1.0"
description = "Python SDK for the Web Crawler gRPC service"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "grpcio-tools>=1.60.0",
    "mypy>=1.8",
    "mypy-protobuf>=3.5",
    "ruff>=0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "crawler.pb.*"
ignore_errors = true

[tool.ruff]
target-version = "py310"
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools.packages.find]
include = ["crawler*"]
